+++ core_java/jdbc/src/test/java/ca/jrvs/apps/stockquote/service/PositionService_IntTest.java
@@ -1,4 +1,4 @@
-package ca.jrvs.apps.stockquote.helper;
+package ca.jrvs.apps.stockquote.service;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -8,8 +8,7 @@ import ca.jrvs.apps.stockquote.dao.PositionDao;
 import ca.jrvs.apps.stockquote.dao.QuoteDao;
 import ca.jrvs.apps.stockquote.dto.Position;
 import ca.jrvs.apps.stockquote.dto.Quote;
-import ca.jrvs.apps.stockquote.service.PositionService;
-import ca.jrvs.apps.stockquote.service.QuoteService;
+import ca.jrvs.apps.stockquote.helper.QuoteHttpHelper;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.Optional;
